requirements.txt
# Flask and API
Flask==3.0.0
flask-cors==4.0.0

# Database clients
pymongo==4.6.1
neo4j==5.15.0

# Data processing and ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
umap-learn==0.5.5

# Spotify API
spotipy==2.23.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
